pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "elasticsearch-mcp-server"
version = "1.0.0"
description = "MCP server for Elasticsearch semantic search (ELSER, E5, hybrid, RRF)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["elasticsearch", "mcp", "semantic-search", "elser", "e5", "vector-search"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.0.0",
    "elasticsearch[async]>=8.12.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
elasticsearch-mcp = "elasticsearch_mcp.__main__:main"
elasticsearch-api = "elasticsearch_mcp.api:main"

[tool.hatch.build.targets.wheel]
packages = ["src/elasticsearch_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
